Extract shared blocking catch-up plumbing into AbstractCatchupSubscriptionModel
The module split duplicated the pass-through lifecycle and storage-config plumbing across
StreamCatchupSubscriptionModel and DcbCatchupSubscriptionModel. Hoist the byte-identical parts
into an abstract base in the DCB-free stream module. The catch-up loops and subscribe routing
stay in the subclasses.
