+++ subscription/util/blocking/stream-catchup-subscription/src/main/java/org/occurrent/subscription/blocking/durable/catchup/StreamCatchupSubscriptionModel.java
@@ -32,7 +32,6 @@ import org.occurrent.subscription.StartAt.SubscriptionModelContext;
 import org.occurrent.subscription.StringBasedSubscriptionPosition;
 import org.occurrent.subscription.SubscriptionFilter;
 import org.occurrent.subscription.SubscriptionPosition;
-import org.occurrent.subscription.api.blocking.DelegatingSubscriptionModel;
 import org.occurrent.subscription.api.blocking.PositionAwareSubscriptionModel;
 import org.occurrent.subscription.api.blocking.Subscription;
 import org.occurrent.subscription.api.blocking.SubscriptionModel;
@@ -45,11 +44,8 @@ import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
 import java.util.Objects;
-import java.util.Optional;
 import java.util.StringJoiner;
 import java.util.concurrent.CompletableFuture;
-import java.util.concurrent.ConcurrentHashMap;
-import java.util.concurrent.ConcurrentMap;
 import java.util.concurrent.Future;
 import java.util.function.Consumer;
 import java.util.function.Function;
@@ -72,14 +68,9 @@ import static org.occurrent.time.internal.RFC3339.RFC_3339_DATE_TIME_FORMATTER;
  * documented on the dispatcher.
  */
 @NullMarked
-public class StreamCatchupSubscriptionModel implements SubscriptionModel, DelegatingSubscriptionModel {
+public class StreamCatchupSubscriptionModel extends AbstractCatchupSubscriptionModel {
 
-    private final PositionAwareSubscriptionModel subscriptionModel;
     private final EventStoreQueries eventStoreQueries;
-    private final CatchupSubscriptionModelConfig config;
-    private final Class<?> subscriptionModelContextType;
-    private final ConcurrentMap<String, Boolean> runningCatchupSubscriptions = new ConcurrentHashMap<>();
-    private volatile boolean shuttingDown = false;
 
     public StreamCatchupSubscriptionModel(PositionAwareSubscriptionModel subscriptionModel, EventStoreQueries eventStoreQueries, CatchupSubscriptionModelConfig config) {
         this(subscriptionModel, eventStoreQueries, config, StreamCatchupSubscriptionModel.class);
@@ -93,10 +84,8 @@ public class StreamCatchupSubscriptionModel implements SubscriptionModel, Delega
      *                                      keeps working regardless of which mode-specific class runs the catch-up.
      */
     public StreamCatchupSubscriptionModel(PositionAwareSubscriptionModel subscriptionModel, EventStoreQueries eventStoreQueries, CatchupSubscriptionModelConfig config, Class<?> subscriptionModelContextType) {
-        this.subscriptionModel = Objects.requireNonNull(subscriptionModel, "subscriptionModel cannot be null");
+        super(subscriptionModel, config, subscriptionModelContextType);
         this.eventStoreQueries = Objects.requireNonNull(eventStoreQueries, "eventStoreQueries cannot be null");
-        this.config = Objects.requireNonNull(config, "config cannot be null");
-        this.subscriptionModelContextType = Objects.requireNonNull(subscriptionModelContextType, "subscriptionModelContextType cannot be null");
     }
 
     /**
@@ -502,80 +491,11 @@ public class StreamCatchupSubscriptionModel implements SubscriptionModel, Delega
         }
     }
 
-    // Reports subscriptionModelContextType (CatchupSubscriptionModel when wrapped by the dispatcher) so a
-    // StartAt.dynamic supplied by a caller that pattern matches on the public dispatcher type keeps working
-    // regardless of which mode-specific class ends up running the catch-up underneath it.
-    SubscriptionModelContext generateSubscriptionModelContext() {
-        return new SubscriptionModelContext(subscriptionModelContextType);
-    }
-
-    @Override
-    public void stop() {
-        getDelegatedSubscriptionModel().stop();
-    }
-
-    @Override
-    public void start(boolean resumeSubscriptionsAutomatically) {
-        getDelegatedSubscriptionModel().start(resumeSubscriptionsAutomatically);
-    }
-
-    @Override
-    public boolean isRunning() {
-        return getDelegatedSubscriptionModel().isRunning();
-    }
-
-    @Override
-    public boolean isRunning(String subscriptionId) {
-        return getDelegatedSubscriptionModel().isRunning(subscriptionId);
-    }
-
-    @Override
-    public boolean isPaused(String subscriptionId) {
-        return getDelegatedSubscriptionModel().isPaused(subscriptionId);
-    }
-
-    @Override
-    public Subscription resumeSubscription(String subscriptionId) {
-        return getDelegatedSubscriptionModel().resumeSubscription(subscriptionId);
-    }
-
-    @Override
-    public void pauseSubscription(String subscriptionId) {
-        getDelegatedSubscriptionModel().pauseSubscription(subscriptionId);
-    }
-
-    /**
-     * Cancel a stream catch-up running for {@code subscriptionId}. A no-op if this class has no catch-up running for
-     * that id (for example because it belongs to the DCB path in a dual-mode dispatcher). Does not touch the shared
-     * live delegate or position storage; the dispatcher owns those since both paths share the same delegate.
-     */
-    public void cancelRunningCatchup(String subscriptionId) {
-        runningCatchupSubscriptions.remove(subscriptionId);
-    }
-
-    /**
-     * Mark this model as shutting down so any in-flight catch-up stops as soon as possible. Does not touch the shared
-     * live delegate; the dispatcher owns that.
-     */
-    public void markShuttingDown() {
-        shuttingDown = true;
-        runningCatchupSubscriptions.clear();
-    }
-
     @Override
     public void cancelSubscription(String subscriptionId) {
         cancelRunningCatchup(subscriptionId);
         subscriptionModel.cancelSubscription(subscriptionId);
         deletePositionFromStorage(subscriptionId);
-    }
-
-    /**
-     * Delete {@code subscriptionId}'s position from the configured position storage, if any. Exposed so the
-     * dispatcher can delete it exactly once when cancelling a subscription that could belong to either mode, since
-     * the position storage config (and the storage instance it wraps) is shared, not owned per mode.
-     */
-    public void deletePositionFromStorage(String subscriptionId) {
-        doIfSubscriptionPositionStorageConfigIs(UseSubscriptionPositionInStorage.class, cfg -> cfg.storage().delete(subscriptionId));
     }
 
     @Override
@@ -617,24 +537,6 @@ public class StreamCatchupSubscriptionModel implements SubscriptionModel, Delega
     private static boolean startsAtBeginningOfTime(StartAt startAt, Class<?> contextType) {
         StartAt start = startAt.get(new SubscriptionModelContext(contextType));
         return start instanceof StartAtSubscriptionPosition position && isBeginningOfTime(position.subscriptionPosition);
-    }
-
-    @Override
-    public SubscriptionModel getDelegatedSubscriptionModel() {
-        return subscriptionModel;
-    }
-
-    private <T, C extends SubscriptionPositionStorageConfig> Optional<T> returnIfSubscriptionPositionStorageConfigIs(Class<C> cls, Function<C, @Nullable T> fn) {
-        if (cls.isInstance(config.subscriptionStorageConfig)) {
-            return Optional.ofNullable(fn.apply(cls.cast(config.subscriptionStorageConfig)));
-        }
-        return Optional.empty();
-    }
-
-    private <C extends SubscriptionPositionStorageConfig> void doIfSubscriptionPositionStorageConfigIs(Class<C> cls, Consumer<C> consumer) {
-        if (cls.isInstance(config.subscriptionStorageConfig)) {
-            consumer.accept(cls.cast(config.subscriptionStorageConfig));
-        }
     }
 
     @Override
